# pkg/dashboard/swaggerdocs/swagger.yaml
...
basePath: /api
definitions:
  archive.Archive:
    $ref: '#/definitions/core.ObjectBase'
  archive.Detail:
    properties:
      kube_object:
        $ref: '#/definitions/core.KubeObjectDesc'
        type: object
    type: object
...
  core.WorkflowDetail:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      entry:
        description: the entry node name
        type: string
      id:
        type: integer
      kube_object:
        $ref: '#/definitions/core.KubeObjectDesc'
        type: object
      name:
        type: string
      namespace:
        type: string
      status:
        type: string
      topology:
        $ref: '#/definitions/core.Topology'
        type: object
      uid:
        type: string
    type: object
  core.WorkflowMeta:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      entry:
        description: the entry node name
        type: string
      id:
        type: integer
      name:
        type: string
      namespace:
        type: string
      status:
        type: string
      uid:
        type: string
    type: object
...
  schedule.Detail:
    properties:
      created_at:
        type: string
      experiment_uids:
        items:
          type: string
        type: array
      kind:
        type: string
      kube_object:
        $ref: '#/definitions/core.KubeObjectDesc'
        type: object
      name:
        type: string
      namespace:
        type: string
      status:
        type: string
      uid:
        type: string
    type: object
  schedule.Schedule:
    properties:
      created_at:
        type: string
      kind:
        type: string
      name:
        type: string
      namespace:
        type: string
      status:
        type: string
      uid:
        type: string
    type: object
...
  utils.APIError:
    properties:
      code:
        type: integer
      full_text:
        type: string
      message:
        type: string
      type:
        type: string
    type: object
  utils.Response:
    properties:
      status:
        type: string
    type: object
...
  v1alpha1.Schedule:
    properties:
      spec:
        $ref: '#/definitions/v1alpha1.ScheduleSpec'
        type: object
      status:
        $ref: '#/definitions/v1alpha1.ScheduleStatus'
        description: +optional
        type: object
    type: object
...
  workflow.StatusResponse:
    properties:
      status:
        type: string
    type: object
info:
  contact:
    name: GitHub Issues
    url: https://github.com/chaos-mesh/chaos-mesh/issues
  description: Swagger for Chaos Mesh Dashboard. If you encounter any problems with
    API, please click on the issues link below to report.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Chaos Mesh Dashboard API
  version: "2.0"
paths:
  /archives:
    delete:
      description: Delete the specified archived experiment.
      parameters:
      - description: uids
        in: query
        name: uids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Delete the specified archived experiment.
      tags:
      - archives
    get:
      description: Get archived chaos experiments.
      parameters:
      - description: namespace
        in: query
        name: namespace
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: kind
        enum:
        - PodChaos
        - IOChaos
        - NetworkChaos
        - TimeChaos
        - KernelChaos
        - StressChaos
        in: query
        name: kind
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/archive.Archive'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get archived chaos experiments.
      tags:
      - archives
  /archives/{uid}:
    delete:
      description: Delete the specified archived experiment.
      parameters:
      - description: uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Delete the specified archived experiment.
      tags:
      - archives
    get:
      description: Get the archived chaos experiment's detail by uid.
      parameters:
      - description: the archive uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/archive.Detail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get an archived chaos experiment.
      tags:
      - archives
  /archives/schedules:
    delete:
      description: Delete the specified archived schedule.
      parameters:
      - description: uids
        in: query
        name: uids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Delete the specified archived schedule.
      tags:
      - archives
    get:
      description: Get archived schedule experiments.
      parameters:
      - description: namespace
        in: query
        name: namespace
        type: string
      - description: name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/archive.Archive'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get archived schedule experiments.
      tags:
      - archives
  /archives/schedules/{uid}:
    delete:
      description: Delete the specified archived schedule.
      parameters:
      - description: uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Delete the specified archived schedule.
      tags:
      - archives
    get:
      description: Get the detail of an archived schedule experiment.
      parameters:
      - description: uid
        in: query
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/archive.Detail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get the detail of an archived schedule experiment.
      tags:
      - archives
  /archives/workflows:
    delete:
      description: Delete the specified archived workflows.
      parameters:
      - description: uids
        in: query
        name: uids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Delete the specified archived workflows.
      tags:
      - archives
    get:
      description: Get archived workflow.
      parameters:
      - description: namespace
        in: query
        name: namespace
        type: string
      - description: name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/archive.Archive'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get archived workflow.
      tags:
      - archives
  /archives/workflows/{uid}:
    delete:
      description: Delete the specified archived workflow.
      parameters:
      - description: uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Delete the specified archived workflow.
      tags:
      - archives
    get:
      description: Get the detail of an archived workflow.
      parameters:
      - description: uid
        in: query
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/archive.Detail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get the detail of an archived workflow.
      tags:
      - archives
...
  /schedules:
    delete:
      description: Batch delete schedules by uids.
      parameters:
      - description: 'the schedule uids, split with comma. Example: ?uids=uid1,uid2'
        in: query
        name: uids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Batch delete schedules.
      tags:
      - schedules
    get:
      description: Get chaos schedules from k8s cluster in real time.
      parameters:
      - description: filter schedules by namespace
        in: query
        name: namespace
        type: string
      - description: filter schedules by name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schedule.Schedule'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: List chaos schedules.
      tags:
      - schedules
    post:
      consumes:
      - application/json
      description: Pass a JSON object to create a new schedule. The schema for JSON
        is the same as the YAML schema for the Kubernetes object.
      parameters:
      - description: the schedule definition
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/v1alpha1.Schedule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha1.Schedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Create a new schedule.
      tags:
      - schedules
    put:
      description: Update a schedule.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1alpha1.Schedule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha1.Schedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Update a schedule.
      tags:
      - schedules
  /schedules/{uid}:
    delete:
      description: Delete the schedule by uid.
      parameters:
      - description: the schedule uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Delete a schedule.
      tags:
      - schedules
    get:
      description: Get the schedule's detail by uid.
      parameters:
      - description: the schedule uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schedule.Detail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get a schedule.
      tags:
      - schedules
...
  /workflows:
    get:
      description: List workflows from Kubernetes cluster.
      parameters:
      - description: namespace, given empty string means list from all namespace
        in: query
        name: namespace
        type: string
      - description: status
        enum:
        - Initializing
        - Running
        - Errored
        - Finished
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.WorkflowMeta'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: List workflows from Kubernetes cluster.
      tags:
      - workflows
  /workflows/{uid}:
    delete:
      description: Delete the specified workflow.
      parameters:
      - description: uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/workflow.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Delete the specified workflow.
      tags:
      - workflows
    get:
      description: Get detailed information about the specified workflow. If that
        object is not existed in kubernetes, it will only return ths persisted data
        in the database.
      parameters:
      - description: uid
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.WorkflowDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get detailed information about the specified workflow.
      tags:
      - workflows
    put:
      description: Update a workflow.
      parameters:
      - description: uid
        in: path
        name: uid
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1alpha1.Workflow'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.WorkflowDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Update a workflow.
      tags:
      - workflows
...
swagger: "2.0"